Fixed fragment not rendering recursively
