+++ src/types.ts
@@ -1,23 +1,8 @@
+import { Fragment, FragmentMarker } from './helpers'
+
 export function optional<T>(obj: T): T | undefined {
   return obj
 }
-
-export class Fragment<T> {
-  readonly typeName: string
-  readonly fields: T
-
-  constructor(typeName: string, fields: T) {
-    this.typeName = typeName
-    this.fields = fields
-  }
-
-  render() {
-    const joinedFields = Object.keys(this.fields).join(' ')
-    return `... on ${this.typeName} { ${joinedFields} }` as any
-  }
-}
-
-export const FragmentMarker = '__fragment__on__'
 
 export function on<T extends {}>(typeName: string, fields: T): Partial<T> {
   const fragment = new Fragment(typeName, fields)
